Limit scope of ServiceInvocationFactory to ingress_grpc crate
