+++ src/common/Cargo.toml
@@ -15,6 +15,5 @@ bytestring = { workspace = true }
 serde = { workspace = true }
 uuid = { workspace = true }
 opentelemetry_api = { workspace = true }
-thiserror = { workspace = true }
 tokio = { workspace = true, default-features = false, features = ["time"]}
 xxhash-rust = { version = "0.8", features = ["xxh3"] }
